moved `Emcee` to externals. But left ParticleStep
